adding reviews

+++ app/src/main/java/com/example/corngrain/data/network/response/Detail.kt
@@ -10,7 +10,7 @@ data class Detail(
     @SerializedName("belongs_to_collection")
     val belongsToCollection: Any,
     @SerializedName("budget")
-    val budget: Int,
+    val budget: Long,
     @SerializedName("genres")
     val genres: List<Genre>,
     @SerializedName("homepage")
@@ -36,7 +36,7 @@ data class Detail(
     @SerializedName("release_date")
     val releaseDate: String,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("runtime")
     val runtime: Int,
     @SerializedName("spoken_languages")
